Migrate schema and data into new layout
This commit updates the project to use a new database schema
structure which divides the schema and data sqlite3 files into the
schema and data directories.

Included are updates to the Makefile, database maintenance scripts
to create, apply patches, and migrate the data into the new layout.

The name of the schema and data files also now follow a new naming
convention.